Added a message center for collecting and dispatching all messages. Messages are passed to a chain of handlers. a new VSLTFMessage is defined. The old one is renamed to VSTLFRealTimeMessage which deriveds from VSTFLMessage. Redirect the real time messages to go through the message center.
(I did not use the old channel since the VSTLFNotificationCenter has
a fixed set of interfaces for each messsage. It means I have to implement the
interfaces for processing realtime messages even they are not interesting.
And everytime a new type message is added I have to rewrite the handlers.
So the realtime messages are dealed by a RealTimeMsgHandler which forwards the messages to the
notification center. New messages will derive from VSTLFMessage)